templates/base.html: fix rel_top calculation in Zola 0.12
For the new wwt-research-kit docs, I see a problem where `rel_top` is
getting too many `..` components. For this site I upgraded from Zola
0.10 to Zola 0.12 for the first time, so I have to imagine that it's
related to something there.

The value of `current_path` which is being used to determine `rel_top`
is `/apiref/` in my example, so I'm guessing that there was a behavior
change where the trailing slash is now causing an extra item to show up
in the for loop. In an attempt to solve this backwards-compatibly, strip
off a trailing slash. This seems to work.